size_t changed for imgno
git-svn-id: file:///home/roberto/Git/svn@7129 1a721e8d-d38c-4e91-ab98-6a0be4d754af

Former-commit-id: 453984e164c66fc314de36def5bd63a9b7df5363

+++ trunk/xmipp/libraries/reconstruction/angular_projection_matching.h
@@ -184,7 +184,7 @@ public:
 
     /** Read current image into memory and translate accoring to
       previous optimal Xoff and Yoff */
-    void getCurrentImage(int imgno, Image<double> &img);
+    void getCurrentImage(size_t imgno, Image<double> &img);
 
 };				
 //@}
